accessibility/tree/checkbox-indeterminate.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <style>
    </style>
</head>
<body>

<input type="checkbox" checked aria-checked="mixed">

<script>
document.querySelector("[type=checkbox]").indeterminate = true;
</script>

<script>
promise_test(async t => {
    assert_tree(
        await test_driver
            .get_accessibility_tree(),
        {
            "nodes": [
                {
                    "backendDOMNodeId": 3,
                    "childIds": [
                        "2"
                    ],
                    "chromeRole": {
                        "type": "internalRole",
                        "value": 144
                    },
                    "ignored": false,
                    "name": {
                        "sources": [
                            {
                                "attribute": "aria-labelledby",
                                "type": "relatedElement"
                            },
                            {
                                "attribute": "aria-label",
                                "type": "attribute"
                            },
                            {
                                "attribute": "aria-label",
                                "superseded": true,
                                "type": "attribute"
                            },
                            {
                                "nativeSource": "title",
                                "type": "relatedElement"
                            },
                            {
                                "attribute": "title",
                                "superseded": true,
                                "type": "attribute"
                            }
                        ],
                        "type": "computedString",
                        "value": ""
                    },
                    "nodeId": "1",
                    "properties": [
                        {
                            "name": "focusable",
                            "value": {
                                "type": "booleanOrUndefined",
                                "value": true
                            }
                        },
                        {
                            "name": "focused",
                            "value": {
                                "type": "booleanOrUndefined",
                                "value": true
                            }
                        }
                    ],
                    "role": {
                        "type": "internalRole",
                        "value": "RootWebArea"
                    }
                },
                {
                    "backendDOMNodeId": 5,
                    "childIds": [
                        "3"
                    ],
                    "chromeRole": {
                        "type": "internalRole",
                        "value": 0
                    },
                    "ignored": true,
                    "ignoredReasons": [
                        {
                            "name": "uninteresting",
                            "value": {
                                "type": "boolean",
                                "value": true
                            }
                        }
                    ],
                    "nodeId": "2",
                    "parentId": "1",
                    "role": {
                        "type": "role",
                        "value": "none"
                    }
                },
                {
                    "backendDOMNodeId": 6,
                    "childIds": [
                        "4"
                    ],
                    "chromeRole": {
                        "type": "internalRole",
                        "value": 88
                    },
                    "ignored": false,
                    "name": {
                        "sources": [
                            {
                                "attribute": "aria-labelledby",
                                "type": "relatedElement"
                            },
                            {
                                "attribute": "aria-label",
                                "type": "attribute"
                            },
                            {
                                "attribute": "title",
                                "type": "attribute"
                            }
                        ],
                        "type": "computedString",
                        "value": ""
                    },
                    "nodeId": "3",
                    "parentId": "2",
                    "properties": [],
                    "role": {
                        "type": "role",
                        "value": "generic"
                    }
                },
                {
                    "backendDOMNodeId": 2,
                    "childIds": [],
                    "chromeRole": {
                        "type": "internalRole",
                        "value": 14
                    },
                    "ignored": false,
                    "name": {
                        "sources": [
                            {
                                "attribute": "aria-labelledby",
                                "type": "relatedElement"
                            },
                            {
                                "attribute": "aria-label",
                                "type": "attribute"
                            },
                            {
                                "nativeSource": "label",
                                "type": "relatedElement"
                            },
                            {
                                "type": "contents"
                            },
                            {
                                "attribute": "title",
                                "type": "attribute"
                            }
                        ],
                        "type": "computedString",
                        "value": ""
                    },
                    "nodeId": "4",
                    "parentId": "3",
                    "properties": [
                        {
                            "name": "invalid",
                            "value": {
                                "type": "token",
                                "value": "false"
                            }
                        },
                        {
                            "name": "focusable",
                            "value": {
                                "type": "booleanOrUndefined",
                                "value": true
                            }
                        },
                        {
                            "name": "checked",
                            "value": {
                                "type": "tristate",
                                "value": "mixed"
                            }
                        }
                    ],
                    "role": {
                        "type": "role",
                        "value": "checkbox"
                    }
                }
            ]
        }
    );
});

function assert_tree(actual, expect) {
    if (typeof expect === "object") {
        assert_equals(typeof actual, "object");
        for (const key in expect) {
            assert_own_property(actual, key);
            assert_tree(actual[key], expect[key]);
        }
    } else {
        assert_equals(actual, expect);
    }
}
</script>

</body>
</html>
